egonet_from_nodelist.py
Function takes a nodelist and generates an edgelist using all possible combinations of nodelist row content WITHOUT checking for appropriateness of data manipulation (i.e. unipartite, bipartite, directed/undirected). Outputs edgelist to csv which should be treated as undirected if being used for network analysis.
This function is useful for hamfisted bipartite to unipartite transformations where nodelist-type data is easiest to obtain. For example, transforming a co-authorship network where authorship is typically presented as a row of delimited authors.